+++ Modules/Intent.Modules.Eventing.Contracts/Templates/IntegrationEventDto/IntegrationEventDtoTemplatePartial.cs
@@ -4,10 +4,12 @@ using System.Linq;
 using Intent.Engine;
 using Intent.Modelers.Eventing.Api;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Api;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.CSharp.TypeResolvers;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Common.Types.Api;
 using Intent.Modules.Constants;
 using Intent.Modules.Eventing.Contracts.Templates.IntegrationEventEnum;
 using Intent.RoslynWeaver.Attributes;
@@ -31,8 +33,22 @@ namespace Intent.Modules.Eventing.Contracts.Templates.IntegrationEventDto
             AddTypeSource(IntegrationEventEnumTemplate.TemplateId);
             AddTypeSource(TemplateRoles.Domain.Enum);
 
+            var classNamespace = Model.InternalElement.Package.Name.ToCSharpNamespace();
+            var extendedNamespace = Model.GetParentFolders().Where(x =>
+            {
+                if (string.IsNullOrWhiteSpace(x.Name))
+                {
+                    return false;
+                }
+
+                return !x.HasFolderOptions() || x.GetFolderOptions().NamespaceProvider();
+            })
+            .Select(x => x.Name);
+
+            var completeNamespace = string.Join(".", classNamespace.Split(".").Concat(extendedNamespace));
+
             CSharpFile = new CSharpFile(
-                    @namespace: Model.InternalElement.Package.Name.ToCSharpNamespace(),
+                    @namespace: completeNamespace,
                     relativeLocation: this.GetFolderPath())
                 .AddClass($"{Model.Name}", @class =>
                 {