+++ Modules/Intent.Modules.Eventing.Contracts/Templates/IntegrationEventMessage/IntegrationEventMessageTemplatePartial.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Intent.Engine;
 using Intent.Modelers.Eventing.Api;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Api;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.CSharp.TypeResolvers;
@@ -11,6 +12,7 @@ using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.TypeResolution;
 using Intent.Modules.Common.Types.Api;
 using Intent.Modules.Constants;
+using Intent.Modules.Eventing.Contracts.Templates.IntegrationEventDto;
 using Intent.Modules.Eventing.Contracts.Templates.IntegrationEventEnum;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
@@ -33,9 +35,23 @@ namespace Intent.Modules.Eventing.Contracts.Templates.IntegrationEventMessage
             AddTypeSource(IntegrationEventEnumTemplate.TemplateId);
             AddTypeSource(TemplateRoles.Domain.Enum);
             AddTypeSource(TemplateRoles.Application.Contracts.Enum);
+            AddTypeSource(IntegrationEventDtoTemplate.TemplateId);
+
+            var classNamespace = Model.InternalElement.Package.Name.ToCSharpNamespace();
+            var extendedNamespace = Model.GetParentFolders().Where(x =>
+            {
+                if (string.IsNullOrWhiteSpace(x.Name))
+                {
+                    return false;
+                }
+
+                return !x.HasFolderOptions() || x.GetFolderOptions().NamespaceProvider();
+            })
+            .Select(x => x.Name);
+            var completeNamespace = string.Join(".", classNamespace.Split(".").Concat(extendedNamespace));
 
             CSharpFile = new CSharpFile(
-                    @namespace: Model.InternalElement.Package.Name.ToCSharpNamespace(),
+                    @namespace: completeNamespace,
                     relativeLocation: this.GetFolderPath())
                 .AddRecord($"{Model.Name.RemoveSuffix("Event")}Event", record =>
                 {
@@ -56,7 +72,7 @@ namespace Intent.Modules.Eventing.Contracts.Templates.IntegrationEventMessage
 
                     foreach (var property in Model.Properties)
                     {
-                        record.AddProperty(GetTypeName(property.TypeReference), property.Name.ToPascalCase(), p =>
+                        record.AddProperty(GetTypeName(property), property.Name.ToPascalCase(), p =>
                         {
                             p.Init();
                             p.RepresentsModel(property);